Added some comments, fixed rotation, removed redundant files
